POWER
Partner Organizations for World Environment and Resources: An ideal
collaboration of responsible businesses that strive to provide quality
products and services of the highest standard whilst harnessing the
values and tradition to sustain existing natural resources for the
wellbeing of surrounding communities. This in 2013-14 gave rise to more
meaningful initiative that have empowered Social Media and its users
automation and real-time connectivity to fullfil fundamental roles;
essential for accelerating global development.